requirements.txt
anyio==3.7.1
azure-core==1.28.0
azure-cosmos==4.4.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==1.10.11
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
